Add API documentation via yuidoc (#122)
* [DOCS] Add API documentation via yuidoc

- add ember-cli-yuidoc and yuidoc-ember-theme as dependencies
- fix documentation in classes

* move functions to separate utility file
