+++ addon/components/sortable-group.js
@@ -7,37 +7,53 @@ const { A, Component, get, set, run } = Ember;
 const a = A;
 const NO_MODEL = {};
 
+/**
+ * @module ember-sortable
+ * @class SortableGroup
+ * @extends Ember.Component
+ */
 export default Component.extend({
   layout: layout,
 
   attributeBindings: ['data-test-selector'],
 
   /**
-    @property direction
-    @type string
-    @default y
-  */
+   * Direction this group can be sorted in. Can be either 'y' or 'x'
+   *
+   * @property direction
+   * @type string
+   * @default y
+   * @public
+   */
   direction: 'y',
 
   /**
-    @property model
-    @type Any
-    @default null
-  */
+   * Optional "group model" to set. If this property is set,
+   * it will be passed in as the first argument to the
+   * {{#crossLink "SortableGroup/onChange:property"}}onChange action{{/crossLink}}
+   *
+   * @property model
+   * @type Any
+   * @default null
+   * @public
+   */
   model: NO_MODEL,
 
   /**
-    @property items
-    @type Ember.NativeArray
-  */
+   * @type Ember.NativeArray
+   * @property items
+   * @public
+   */
   items: computed(() => a()),
 
   /**
-    Position for the first item.
-    If spacing is present, first item's position will have to change as well.
-    @property itemPosition
-    @type Number
-  */
+   * Position for the first item.
+   * If spacing is present, first item's position will have to change as well.
+   *
+   * @property itemPosition
+   * @type Number
+   * @public
+   */
   itemPosition: computed(function() {
     let direction = this.get('direction');
 
@@ -45,9 +61,12 @@ export default Component.extend({
   }).volatile(),
 
   /**
-    @property sortedItems
-    @type Array
-  */
+   * List of items sorted by direction.
+   *
+   * @property sortedItems
+   * @type Array
+   * @public
+   */
   sortedItems: computed(function() {
     let items = a(this.get('items'));
     let direction = this.get('direction');
@@ -56,37 +75,45 @@ export default Component.extend({
   }).volatile(),
 
   /**
-    Register an item with this group.
-    @method registerItem
-    @param {SortableItem} [item]
-  */
+   * Register an item with this group.
+   *
+   * @method registerItem
+   * @param {SortableItem} item
+   * @public
+   */
   registerItem(item) {
     this.get('items').addObject(item);
   },
 
   /**
-    De-register an item with this group.
-    @method deregisterItem
-    @param {SortableItem} [item]
-  */
+   * De-register an item with this group.
+   *
+   * @method deregisterItem
+   * @param {SortableItem} item
+   * @public
+   */
   deregisterItem(item) {
     this.get('items').removeObject(item);
   },
 
   /**
-    Prepare for sorting.
-    Main purpose is to stash the current itemPosition so
-    we don’t incur expensive re-layouts.
-    @method prepare
-  */
+   * Prepare for sorting.
+   * Main purpose is to stash the current itemPosition so
+   * we don’t incur expensive re-layouts.
+   *
+   * @method prepare
+   * @public
+   */
   prepare() {
     this._itemPosition = this.get('itemPosition');
   },
 
   /**
-    Update item positions (relatively to the first element position).
-    @method update
-  */
+   * Update item positions (relative to the first element position).
+   *
+   * @method update
+   * @public
+   */
   update() {
     let sortedItems = this.get('sortedItems');
     // Position of the first element
@@ -122,8 +149,11 @@ export default Component.extend({
   },
 
   /**
-    @method commit
-  */
+   * Commit the state of dragged items
+   *
+   * @method commit
+   * @public
+   */
   commit() {
     let items = this.get('sortedItems');
     let groupModel = this.get('model');
@@ -158,4 +188,12 @@ export default Component.extend({
       invokeAction(this, 'onChange', itemModels, draggedModel);
     }
   }
+
+  /**
+   * String name or action function to call whenever the sort order changes.
+   *
+   * @event onChange
+   * @type Action|String
+   * @public
+   */
 });
