+++ addon/helpers/reorder.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 
-/**
+/*
   In tests, the dummy app is rendered at half size.
   To avoid rounding errors, we must therefore double
   the overshoot.
@@ -8,28 +8,27 @@ import Ember from 'ember';
 const OVERSHOOT = 2;
 
 /**
-  Reorders elements to the specified state.
+ * Reorders elements to the specified state.
+ *
+ * Example:
 
-  Examples
-
-      reorder(
-        'mouse',
-        '.some-list li',
-        '[data-id="66278893"]',
-        '[data-id="66278894"]',
-        '[data-id="66278892"]'
-      );
-
-  @method reorder
-  @param {'mouse'|'touch'} [mode]
-    event mode
-  @param {String} [itemSelector]
-    selector for all items
-  @param {...String} [resultSelectors]
-    selectors for the resultant order
-  @return {Promise}
-*/
-
+```js
+reorder(
+  'mouse',
+  '.some-list li',
+  '[data-id="66278893"]',
+  '[data-id="66278894"]',
+  '[data-id="66278892"]'
+);
+```
+ * @method reorder
+ * @public
+ * @param {Ember.Application} app
+ * @param {'mouse'|'touch'} mode event mode
+ * @param {String} itemSelector selector for all items
+ * @param {String} [...resultSelectors] selectors for the resultant order
+ * @return {Promise}
+ */
 export function reorder(app, mode, itemSelector, ...resultSelectors) {
   const {
     andThen,
