+++ app/components/sortable-item.js
@@ -1,3 +1,1 @@
-import sortableItem from 'ember-sortable/components/sortable-item';
-
-export default sortableItem;
+export {default} from 'ember-sortable/components/sortable-item';
